Fix a couple errors.
1) Timing was way off! Stupid mistake on my side, 2MH is 2
million cycles per second, not 2 thousand!

2) No more bigger screen than necessary, let's add the fancy
stuff later.

3) The external shift machine was overflowing (of course, it
was shifting up to 8 bits to the right in a 8 bits int).

4) The DCX instruction was overflowing: When subtracting 0x1
from 0x0, it was panicing instead of giving 0xff as a result.

+++ 8080/emulator/src/lib/io_devices/external_shift.rs
@@ -62,7 +62,7 @@ pub struct ExternalShiftReader {
 impl InputDevice for ExternalShiftReader {
     fn read(&mut self) -> u8 {
         let v = ((self.shift1.get() as u16) << 8) as u8 | self.shift0.get();
-        ((v >> (8-self.shift_offset.get())) & 0xff)
+        (((v as u16) >> (8-self.shift_offset.get())) & 0xff) as u8
     }
 }
